Decompose big UserPreferences for reuse.
Refactor because I want to reuse configuration via bootstrap vue modal in user preferences pattern for selecting object store default preferences.

```
pytest lib/galaxy_test/selenium/test_personal_information.py::DeleteCurrentAccountTestCase
```
