+++ client/src/components/User/user-styles.scss
@@ -1,0 +1,6 @@
+.pref-content {
+    width: calc(100% - 3rem);
+}
+.pref-icon {
+    width: 3rem;
+}
